Rename Pagination to PaginatedTests
